requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
langchain-community
langchain-core
pypdf
groq
pypdf2
pytesseract
Pillow
python-dotenv
chromadb
pydantic
sentence-transformers
jinja2